搜索页

+++ app/src/main/java/com/pipnet/wallenews/bean/TopicsBean.java
@@ -8,13 +8,13 @@ public class TopicsBean {
     public int contentCount;
     public long createTime;
     public int followerCount;
-    public int id;
+    public long id;
     public boolean ifFollowed;
     public String image;
     public String introduction;
     public int isPrivate;
     public int labelId;
-    public int latestContentId;
+    public long latestContentId;
     public String latestContentSubTitle;
     public String latestContentTitle;
     public String name;
